# Prefix caching optimization 