Removed not needed attempt to restore cached blocks during block allocation, as the chance that newly generated sequence will be found in cached blocks is too small.

Time comparison on Meta-Llama-3-8B, 10 chat iterations:

Master:
Total execution time: 770.27 seconds. 
Avg time to first token: 1909.4 ms

This branch: 
Total execution time: 767.14 seconds. 
Avg time to first token: 1907.2 ms